+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Stock Move: select stock move with current invoice canceled in multiple sale and purchase invoicing form
 - Set trading name automatically when creating or merging purchase order
 - Fix bad domain on permission assistant's user field.
+- Fix interco invoice, sale and purchase order generation.
 
 ## [5.0.0-rc2] - 2018-05-09
 ## Improvements